TAS-333 : Lots of changes so that adding the reference allele to an HDF5 genotype table will work "right"
I ran existing unit tests where possible and they all passed. If I messed something up, we can always revert. This is an initial stab at this which could be refined in many ways. For example, population of a PositionList with reference alleles should probably be moved to a PositionListUtils class.